refactor: Move pathsToGlobPatterns to find-util and remove old-find-util
- Move pathsToGlobPatterns function from old-find-util to find-util
- Update imports in createLinter.ts to use pathsToGlobPatterns from find-util
- Remove old-find-util.ts and its test file as they are no longer needed
- Update createIgnorePatterns to handle missing ignore files gracefully
- Remove test case for explicit ignore file error handling since CLI always provides a default path
- Maintain backward compatibility for CLI usage where ignore files may not exist

This completes the migration to the new find-util system while preserving existing behavior.
